Add model files

+++ Pix/CmdSetViewport.cpp
@@ -1,7 +1,6 @@
 #include "CmdSetViewport.h"
 #include "Viewport.h"
 
-
 bool CmdSetViewport::Execute(const std::vector<std::string>& params)
 {
 	if (params.size() < 4)
@@ -9,10 +8,10 @@ bool CmdSetViewport::Execute(const std::vector<std::string>& params)
 		return false;
 	}
 
-	float x = stof(params[0]);
-	float y = stof(params[1]);
-	float width = stof(params[2]);
-	float height = stof(params[3]);
+	int x = std::stof(params[0]);
+	int y = std::stof(params[1]);
+	int width = std::stof(params[2]);
+	int height = std::stof(params[3]);
 
 	bool show = false;
